test (LottoNumbers) : 런타임 예외 대신 커스텀 예외로 변경
